Actually fixed replay framerate independence. Also introduced hasFinishedSettingFrame.
In the ProcessLogFile() function in Replay.cs, we now yield return new
WaitForFixedUpdate(), instead of simply waiting a frame.

hasFinishedSettingFrame should be set to true when everything for a
given frame has been put into position in the ProcessLogFile()
function. Works for just the avatar, will have to change for multiple
objects.